requirements.txt
asgiref==3.7.2
Django==5.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pytest==7.4.4
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
